requirements.txt
psycopg2
Django~=5.1.6
django-model-utils~=5.0.0
django-environ
requests
uwsgi~=2.0.23
PyMySQL~=1.1.1
djangorestframework~=3.15.2
m3u-parser~=0.4.1
django-cors-headers~=4.7.0
psutil~=7.0.0